Do not use absolute metrics paths.
A few places used absolute path starting with '/chrome/infra/' for
metrics reporting. This is not supported by ts_mon (at present). It
prefixes the given path with the (internal?) prefix '/chrome/infra'
irrespective of whether it is an absolute path.

So, change over to relative paths for now.

BUG=chromium:687712
TEST=None

Change-Id: I2cd1e881de8c4ea3458ebe1799337ef885fd77ff
Reviewed-on: https://chromium-review.googlesource.com/435470
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>