firmware_LockedMe: support IFWI based intel devices
The apollolake SoC doesn't have/use a ME region proper. That
binary is embedded within another data structure that lives
in the BIOS region. However, there is still a descriptor region
that needs to be properly locked down along with the ME's
writable data region which needs to be checked. This CL uses
the fmap's region names for determining if it is an IFWI-based
platform.

BUG=chrome-os-partner:62177
BRANCH=reef
TEST=test passes on reef with locked down ME config

Change-Id: Ic74fc9d0f3b9c22d6eb102c5ec65e05815c24a5b
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/438836
Reviewed-by: Furquan Shaikh <furquan@chromium.org>