[autotest] Return empty str for missing board, os labels
We used to return None from HostInfo for board and os values when the
corresponding labels are missing. OTOH, if the corresponding label
existed, but had no value associated with it (e.g, 'os:'), we would
return ''. In practice, these two cases are identical.

BUG=chromium:678430
TEST=unittests

Change-Id: I3adfdc1d498901ec7f98dc3d15ad992b8839f550
Reviewed-on: https://chromium-review.googlesource.com/441905
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>