[autotest] Use HostInfo to update version labels
This CL slightly changes the semantics of clear_version_labels: All
versions labels are cleared now instead of just the type of version
recognized by the corresponding host class. This was always the intended
behaviour.

This CL also changes the semantics from afe_utils.add_version_label in
that if the version prefix already exists, it is updated in place. This
is the right thing to do because adding a version label to the end of
the list when an earlier label exists has no effect (the first one takes
precedence).

BUG=chromium:678430
TEST=(1) (new) unittests
     (2) in-lab hwtests via moblab

Change-Id: I3939e04953a1dd2ac0cf09f61462128410c0db06
Reviewed-on: https://chromium-review.googlesource.com/517772
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>