tko parser: fix variable scoping
The variable ignored_lines was set to the empty list
in print_and_reset_ignored_lines().  This made the python
parser treat it as local to the function instead of using
the variable defined in the outer scope.  Removing the
assignment is probably the best solution in this case.

BUG=chromium:693610
TEST=none

Change-Id: I5a388855b2d808075e7e265fd1dababa4f082aec
Reviewed-on: https://chromium-review.googlesource.com/444100
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>