Add packager.py option to ONLY tarball and MD5
This is a first of a series of patches to move the work of creating
auto-test tarballs and MD5 sums out of the pkg_postinst stage of the
autotest.eclass and into the src_compile stage.

The motivations for doing this are detailed in the bug, but in summary,
we can avoid computing tarballs and MD5sums for packages which are being
downloaded as binaries.

This patch adds functionality to packager.py to compute and save the
tarball and MD5sum of a test into a directory specified by the caller.

BUG=chromium:285856
TEST=Run 'emerge-veyron_minnie chromeos-base/autotest-tests-wimax' locally
CQ-DEPEND=CL:446976

Change-Id: I9d943057938a6bbbced624353205dad745eaa695
Signed-off-by: Prashant Malani <pmalani@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/446026
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>