firmware_Mosys: Handle 'pd info' not being implemented
The top level mosys 'pd' command is being used for TCPC mux
information with 'pd port X' in addition to 'pd info' that is
used today for printing information about the Google EC based
USB PD chip.

This means that the 'pd' command will start showing up on
devices that do not support 'pd info' so we should also check
for the faft config item that indicates this type of usbpd
chip is present before assuming that 'pd info' will work.

BUG=b:35585876
TEST=pass firmware_Mosys test on eve

Change-Id: I7eea836b5b88f28fc41964e0ccb51ce566d83345
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/448106
Reviewed-by: danny chan <dchan@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>