[autotest] check the num of required DUTs in the bvt pool
DUTs in pool:bvt is used in test_push. Currently we check all the
DUTs in the push server to ensure we have the minimum number of DUTs in
the Ready state. However, not all the DUTs are in pool:bvt. Enforce the
check to only check the DUTs in pool:bvt, and improve the error message
to explicitly mention the DUTs in pool:bvt are in shortage.

BUG=chromium:696882
TEST=unittest and also test in testing server

Change-Id: I34711b3f8315ca2a24d2e9162232b052251e058b
Reviewed-on: https://chromium-review.googlesource.com/449873
Commit-Ready: Shuqian Zhao <shuqianz@chromium.org>
Tested-by: Shuqian Zhao <shuqianz@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>