add test to stress cr50 usb
We need a way to stress the usb connection between the AP and the Cr50
USB. This adds a test to stress that connection. By default it won't
exit on the first usb failure it will run through all of the iterations.
There are a couple of Cr50 issues with USB and we want to see if we are
making improvements in the percent of usb interactions that failed.

BUG=b:35774906
BRANCH=none
TEST=Run the test on a board with 0.0.16 cr50 firmware.

Change-Id: I7a2d320751a70e0b8c7806df448342b55a7aa716
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/450901
Reviewed-by: Wai-Hong Tam <waihong@google.com>