power_SuspendStress: stay awake up to 8 seconds in between S2Rs
The current test runs for 4 hours, staying awake between 0 and 3 seconds
between S2R attempts. This doesn't seem long enough to guarantee we'll
keep an SSH connection alive. Particularly, when using the default
autotest parameters (ServerAliveInterval=900, ServerAliveCountMax=3),
then we often can see a 4 hour test have its SSH session terminated at
about 45 minutes.

One way to combat this is to increase the window of values in which we
select a random "resume time"; if we wait up to 8 seconds, then we will
occasionally give the DUT enough time to service the network properly.

BUG=b:35648179
TEST=test_that power_SuspendStress.4hours

Change-Id: I33d4a811f9bd9b2c4df27071534a0d5bce1335cc
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/450785
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>