Add support for cr50 uart automation
This change refactors the Chrome EC and USBPD support so that there is a
console class that takes in the prefix used by the uart cmd, regexp, and
multicmd controls. There was a lot of code being reused by ChromeUSBPD
and ChromeEC. This makes it easier to add new console devices.

This change also adds support for using the cr50 console.

CQ-DEPEND=CL:453924
BUG=b:35647854
BRANCH=none
TEST=none

Change-Id: I7148f7d512f8013b77643c4f000f4cb65c6d18ff
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/454261
Reviewed-by: Wai-Hong Tam <waihong@google.com>