Modify packager to use pre-computed md5 and tarballs.
Modify the tar_package invocations and compute_checksum calls to only
compute the values if pre-existing ones don't exist. When the script
packager script is run on tests with the '-a tar_only' option. These
will be pre-computed and kept as part of the prebuilt package. So, if
only pre-builts are being fetched, we won't need to recalculate these.

Also, this patch fixes a minor error in the naming of the package's
checksum file.

BUG=chromium:285856
TEST=Run 'emerge-veyron_minnie chromeos-base/autotest-tests-wimax' locally

Change-Id: I344334dabffca429b968c340e7ae24f8c962657a
Reviewed-on: https://chromium-review.googlesource.com/455550
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>