add test to verify cr50 deep sleep
This test counts the number of times cr50 enters deep sleep during a
power_SuspendStress test. If suspend_iterations is not set, then the
test will run power_SuspendStress for the specified duration and then
report the deep sleep count. If suspend_iterations is set, it runs
power_SuspendStress for that number of cycles and verifies the deep
sleep count is equal to that number of cycles once power_SuspendStress
is complete. Cr50 should enter deep sleep every cycle, because the
minimum suspend and resume times are set so the cycle takes longer
than the cr50 sleep delay.

Servo V4 cannot charge the DUT right now, so this test can only be run
for a couple of hours. The regular power_SuspendStress can be run to
verify deep sleep resume more thuroughly. This is just to confirm we are
not getting extra deep sleep wakeups.

BUG=b:35835218
BRANCH=none
TEST=test_that firmware_Cr50DeepSleepStress -b reef $DUT_IP

Change-Id: I7880b338aa6f19d8b5c29317cbdc35b3c383f35a
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455586
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>