[autotest] Change the Servo ssh verifier tag to 'servo_ssh'
The verifier for ssh access to a Servo host or to a DUT was named
'ssh' for both.  This created ambiguity for determining whether a
given ssh failure was for the Servo host or the DUT.  This is
particularly a problem for automated parsing of logs, which need to
make the distinction.

This renames the Servo verifier to 'servo_ssh', so as to remove
the ambiguity.

BUG=None
TEST=Run repair in a local instance.

Change-Id: I72b1653b866f343dda7caf71bcdf955d7ec47c21
Reviewed-on: https://chromium-review.googlesource.com/458701
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>