autoupdate_EndToEndTest: Reinstall stateful if DUT was powerwashed when installing source
When installing an old image, the DUT may be powerwashed: it happens
when EXT4 crypto is enabled but an older image can not handle it.

When it happens, reinstalled stateful.tgz to be able to continue the
test. The signal used for powerwash is the presence of rsync, only
installed in /mnt/stateful.

BUG=chromium:689105
TEST=Enable encrypt on a samus (tune2fs -O encrypt /dev/sda1)
Do an end to end test to FSI R41 6680.57.0 and back to R59 latest.

Change-Id: Iaea3691bee0ee52456e55f2876d55dcb055d8359
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/459125
Reviewed-by: Xixuan Wu <xixuan@chromium.org>