autotest: update test to expect INET_DIAG to be on
In order to support SOCK_DESTROY the INET_DIAG configuration option is
going to be turned on for all kernel versions >= 3.18. This change
updates the test so that it no longer checks for INET_DIAG being turned
off for kernel versions >= 3.18. It still checks it for kernel versions
< 3.18.

BUG=706213
TEST=Run the test

Change-Id: I1c12eb124641250d7b20260dec5170aac0ff08bb
Reviewed-on: https://chromium-review.googlesource.com/461275
Commit-Ready: Abhishek Bhardwaj <abhishekbh@google.com>
Tested-by: Abhishek Bhardwaj <abhishekbh@google.com>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>