[autotest] Add FileStore for HostInfo
File backed store will be used to pass HostInfo in-and-out of the
autoserv process launched for a test, and between various autoserv
processes launched for the tests.
The backing file is file-locked to make sure that mutiple autoserv
processes can safely access the data.

BUG=chromium:678430
TEST=(new) unittests

Change-Id: I831367f0cfc13fc132253eee9cd8d41112877301
Reviewed-on: https://chromium-review.googlesource.com/462385
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>