mark chrome_cr50 ccd commands
Some devices have the cr50 console connected to the servo header, so
they don't need to rely on CCD to send cr50 console commands. Cr50 tests
are written to be used with CCD or without. Having ccd enabled on cr50
changes the sleep behavior, so some of the tests enable or disable CCD
at different points. These commands are unnecessary on devices that are
using the servo header instead of suzyq.

This change adds a decorator for the cr50 functions that should only be
used on tests getting the cr50 console through ccd. The wrapper will
ignore the function call if the test is not using ccd.

BUG=none
BRANCH=none
TEST=run firmware_Cr50DeepSleepStress on poppy and verify ccd_enable and
disable are ignored, but the idle calls still work.

Change-Id: I99fba44d3e58158dd7ea792cf85a9725fa69e35d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/463869
Reviewed-by: Wai-Hong Tam <waihong@google.com>