wake up cr50 before sending commands
If the the servo header is being used to access the Cr50 console instead
of CCD, then it is possible that Cr50 is in regular sleep whenever we
try to send a console command. Cr50 will drop the first character as it
wakes up. This change will send '\n' before trying to send the real
console command to make sure cr50 is awake.

BUG=none
BRANCH=none
TEST=run firmware_Cr50DeepSleepStress on poppy

Change-Id: If0dd781263a12e7e0797ff3452197f5a2898e7d7
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/463870