[autotest] Deployment should use existing servo port.
If a deploy with servo v4 fails on a dut after the servo host and port get saved
into the afe for that particular dut, subsequent deploys for that same
dut will grab new unused ports instead of using the port already
reserved for it.  That seems senseless so let's use the port already
chosen for that dut.  A sanity check is done to make sure the port is
reserved on the same servo host.

BUG=None
TEST=deployed a dut that already had an entry in the AFE and used the
same port.

Change-Id: Ia7dce2249a076ad8d70322677158a7650007479c
Reviewed-on: https://chromium-review.googlesource.com/463788
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>