bluetooth_adapter_tests: wait for condition
Some bluetooth adapter tests need to issue commands and then to
query the adapter properties to see the result. We previous simply
wait for a certain period of time before querying the properties.
This turns out to be rather flaky as different platforms may take
different lengths of time to complete the commands.

In this patch, we add a _wait_for_condition() method to wait for
a function to become true with a specified timeout. The function
would be executed repeated with the specified sleep_interval until
its returned value becomes true or until the timeout expires. This
makes those methods become very reliable.

BUG=chromium:708915
TEST=None. Will be tested in next patch.

Change-Id: If2de9c7dc3ae4186994987230d36ca31eba77021
Reviewed-on: https://chromium-review.googlesource.com/469412
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>