dbus_send: correctly parse new dbus message header
dbus-1.10.12 has changed the message header a bit. Noticeably, it now
includes an optional "time=" field, and a mandatory "serial=" field.
Also, "dest=" is changed to "destination="

https://cgit.freedesktop.org/dbus/dbus/tree/tools/dbus-print-message.c?id=dbus-1.10.12#n545
https://cgit.freedesktop.org/dbus/dbus/tree/tools/dbus-print-message.c?id=dbus-1.10.12#n572

For easier transition, the regex for older header format is kept. That
way this CL can go in before the dbus upgrade, and we could delete it
easily when that happens.

BUG=chromium:704365, b:34841890
TEST=test_that -b x86-generic ... peerd_DiscoverServices peerd_AdvertiseServices

Change-Id: I032706f681537dbe880d358b8c832a86c8ce3a35
Reviewed-on: https://chromium-review.googlesource.com/465546
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>