autoupdate_EndToEndTest: Enable devserver when devserver is set
When a device has been powerwash, we need to reinstall its stateful
partition. We can not invoke the DUT stateful_update script, it has not
been removed. We need to rely on autotest server code, that requires the
devserver to run.
Start the devserver when omaha_host is specified.
It uses the URL from payload or stateful url to know which files to
stage.

BUG=chromium:689105
TEST=In production.
._update_via_test_payloads is called from:
- _install_source_version (devserver_hostname set:
  update payload and stateful using auotest code and devserver.
- finalize_update: stateful partition updated with DUT script.

Change-Id: I9645a2974c3e8f97b3710fbbb2690a10400e26a6
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/465489
Reviewed-by: Xixuan Wu <xixuan@chromium.org>