add test to verify cr50 unlock
Enable the cr50 lock, start the unlock process, and press the power
button a couple of times a second until the cr50 lock is disabled.

This test can only be run on devices that have the cr50 console through
servo. Because it needs access to the cr50 console and control of the
power button.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ibf1cb362c64a6e618775a9eb2575c3d801873e4d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/469130
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>