Add profiler and dep support for tarball & md5 packager precalculation.
Just like tests, profiler and deps tarballs and checksums are also
precalcualted now, during the compilation phase. That way, if the
prebuilt checksum is identical to what is there in the
packages.checksum, we don't need to recompile these. Plus, the tarball
and checksum doesn't need to be recalculated for the prebuilt.

BUG=chromium:285856
TEST=Run 'emerge-veyron_minnie autotest-tests-wimax autotest-deps' locally
CQ-DEPEND=CL:469186

Change-Id: I12fdc296f94d6ecea14d721dbf419c5b14b763a9
Reviewed-on: https://chromium-review.googlesource.com/469146
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>