Cr50Update: raise consistent error if the state isn't restored
Raise 'Could not restore the original image' when the test did not
restore the original image. It is better if we raise the same error
every time we don't restore the image. If we don't restore the original
cr50 image, something may need to be done to restore the device to the
original state.

BUG=none
BRANCH=none
TEST=none

Change-Id: I43407b8cb9471e8f7fe34c979c36d6d8f002713d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/469070
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>