firmware_Cr50Update: handle ignored posted resets
For cr50 images 0.0.19 and beyond cr50 rebooting relies on the firmware
to enable the invalid RW header. On each boot the firmware will try to
enable any invalidated RW headers and if there is an invalid one Cr50
will enable it, wait 1000ms and then reboot. Cr50 will no longer reboot
on its own after usb_update is run.

The support for enabling the invalid cr50 RW header is not in firmware
yet. This changes firmware_Cr50Update so the test not care if cr50 does
not reboot immediately after usb_update is run for rollback updates. We
force the rollback ourselves and reboot, so it is ok if cr50 does not
reboot immediately on its own.

BUG=none
BRANCH=none
TEST=build a cr50 dbg image from TOT and then use it as a 'dev_image'
while running the test.

Change-Id: Ie2acafb4e163ba2697cd4e1421f0f37daf9f2758
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/469075
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Tyler Reid <twreid@google.com>