firmware_Cr50Update: change rollback header checks
The header enable/disable changes when the headers can be read. There
may be a valid header in the inactive RW slot, but 'ver' may show it as
an error. This changes firmware_Cr50Update to stop checking that the
version in the inactive RW slot is equal to the target version. It will
attempt the rollback and if the inactive slot is actually corrupt then
Cr50 will choose the dev image again, so it is fine to attempt a
rollback when the other image is corrupt.

To validate the rollback, it will check that we did change partitions
and cr50 is running the right version.

BUG=none
BRANCH=none
TEST=firmware_Cr50Update.erase_nvmem

Change-Id: I220877255e43ebdc293f86a7473d0015435cb4d9
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/471931
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>