network_WlanDriver: drop fixed interface name list; ask shill instead
We really shouldn't be keeping a fixed list of potential device names.
It's more flexible and easier to just ask shill anyway, so let's do
that.

This incidentally fixes some quirks of the 'interface' API, which caused
us to dump a spurious log message when checking for the description:

04/12 14:41:50.819 ERROR|         interface:0171| Device description not supported on non-wifi interface: wlan0.

This also stops pretending that the test actually handles multiple
interfaces properly.

BUG=none
TEST=`test_that ... network_WlanDriver` with Wifi enabled, and with Wifi
     module forcibly unloaded (to see 'TestNA')

Change-Id: I33eb4eff035d31d462f339d61cd119b9335211f2
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/481439
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>