Updated enterprise_CFM_USBPeripheralHotplugDetect test to detect crashes.
Check for crash function looks for crashes in '/var/spool/', '/home/chronos/'
and /home/chronos/u-*'. This test looks for crashes before and after
plugging in USB peripherals. It also checks for crash after hangouts session
is started. When the crashes are detected, the test does not fail. It only
raises error messaage to alert so that crashes can be identified and processed
in later stage manually for its root cause.

BUG=chromium:709134
TEST=Tested on Guado.

Change-Id: I21266e32069d0adf49b106e756af6cd7701ed343
Reviewed-on: https://chromium-review.googlesource.com/481048
Commit-Ready: Dinesh Kumar Sunkara <dsunkara@chromium.org>
Tested-by: Dinesh Kumar Sunkara <dsunkara@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>