Revert "[autotest] Use AfeHostInfoStore to obtain host information ..."
Something in adb_host pulled in a depepdency on PIL library which is
breaking chromeos-admin's test_push task. This CL added an implicit
import to server/hosts/__init__.py which lead to this chain of imports.
Try to remove this to see if the dependency disappears.

Note that the CL itself was not incorrect. We do want to use HostInfo
for test_push. The fact that chromeos-admin's server_management_lib
imports test_push is broken.

BUG=None
TEST=None

This reverts commit c0e976d011d55fa6f00c67fd65d24e6e2646fd6d.

Change-Id: I7fd3cd55e46c41ffdf944907211fe9f0c7d6facb
Reviewed-on: https://chromium-review.googlesource.com/484627
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>