[autotest] Adapt suite_enumerator.py to changes in suite.py
Changes in server/cros/dynamic_suite/suite.py broke the feature
in suite_enumerator.py that prints the tests in a suite. This patch
fixes that, calls create_fs_getter and list_all_suites outside of
the Suite class (calling in the class is deprecated), and removes
checks for experimental tests, which are no longer supported.

BUG=b/36296378,chromium:710923
TEST=./suite_enumerator.py -a ../ graphics

Change-Id: Ic11249d4b52bafb3a7656f2d51d1921908d5062d
Signed-off-by: Mike Mason <michael.w.mason@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/486219
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Michael W Mason <michael.w.mason@intel.com>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>