[autotest] Block access to AFE's admin pages.
The admin pages for the AFE are backed by django's admin module. The
version of django we use is known to have memory leaks. This causes
repeated accesses to the admin page to result in memory issues on the
server.

Any use of admin pages should be possible from the command line (atest
and friends). moblab plans to have its own UI for this task as well.

BUG=chromium:702302
TEST=Local AFE loads properly, doesn't allow admin access.

Change-Id: Ifd4c778565c0d17b32cdfe5b677484a12e3f8b8b
Reviewed-on: https://chromium-review.googlesource.com/486967
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>