[autotest] Block '*.core' from test results.
When Chrome crashes, we care about the minidump files, but not the
core files.  The core files are large, and expensive to include in
test results, so we want to exclude them.

There's a config option that allows selecting whether core files
are uploaded in sysinfo for test results.  By default, that option
is true.  This changes the option to be false in global_config.ini.
This also re-enables code to include /var/spool/crash, since with
core files excluded, that directory should be safe to include.

BUG=chromium:715228
TEST=Run tests against at DUT with dumps; see non-*.core files collected

Change-Id: I4be856c45df1456893347cf3a48179c92d5f8330
Reviewed-on: https://chromium-review.googlesource.com/486968
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>