cros_firmware: Support the new "BIOS (RW)" field.
chromeos-firmwareupdate was recently changed to add some "RW" fields.
However, those fields now refer to the RO version of the firmware and
the "(RW)" fields refer to (as the name implies) the RW verison of the
firmware.

This commit changes the method of retrieving the RW firmware version to
check first for the "BIOS (RW) version" field and then fallback to the
previous method of just "BIOS version" if the field is not found.

BUG=chromium:714786
TEST=Take `chromeos-firmwareupdate -V` output from a caroline image and
a snappy image. Exercise the code and verify that it's able to extract
the correct BIOS firmware versions.

Change-Id: I653ba9f7ed50d287cc29c2dcce7ce599f60c01d6
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/487868
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>