[moblab] Call new upstart command that correctly restarts apache.
restart moblab-apache-init does not cleanly or correctly restart
apache, call a new upstart config that calls the correct command
to restart apache.

Replace the use of the boto library with a gsutil command.

Use the same path to the gsutil command that the devserver uses.

BUG=chromium:714294
TEST=locally on moblab, unittests

Change-Id: I0befe734d2c6130dadcfdaa5984aec23bba07807
Reviewed-on: https://chromium-review.googlesource.com/497169
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>