cheets_CTSHelper.stress: add to all closing suites.
Currently wmatrix shows that Chrome login and Android start is stable
across all boards and several builds. We would like this to stay this
way. cheets_CTSHelper.stress is sharded and starts Android 100 times.
So we move the test to arc-bvt-cq, bvt-cq and bvt-inline. While this
looks somewhat redundant, it is not as it runs in different queues
and different boards. Even where there is redundancy it just helps
surfacing start flakes more reliably.

Of course this test will cause gardeners and constables some grief
when at times they won't be able to uprev Chrome or the container.
I think this is better than the alternative, namely shutting the lab
down to most Android testing for days or weeks at a time.

BUG=None.
TEST=cq will test.

Change-Id: I0266b441b5e53f6b1a39cafebb47ae60603d6c98
Reviewed-on: https://chromium-review.googlesource.com/491046
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>