autotest: Move delay from power_RPMTest to power_CheckAC
This moves the delay to let the AC status settle into power_CheckAC.
This allows a polling loop to be used to make the test faster. The
max time is double (10 seconds) what the standard wait time was
before, but the AC status should change well before that in most
cases.

BUG=None
TEST=power_RPMTest

Change-Id: I2432cdb4c8700efe6849cabe52663e25b8b267cf
Reviewed-on: https://chromium-review.googlesource.com/492309
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>