firmware_Cr50Update: get images from gs://
This change adds support for fetching the release and dbg images from
google storage. If a local path is not specified for the image then it
will be retrieved from google storage.

The prod images are in gs://chromeos-localmirror/distfiles/ and the dev
signed debug images are stored in.
gs://chromeos-localmirror-private/distfiles/chromeos-cr50-debug. The
version string is needed to select the correct prod image. For the debug
image, the test will get the devids from cr50 and then find the right
debug image.

BUG=none
BRANCH=none
TEST=test_that $DUT_IP -b reef firmware_Cr50Update.erase_nvmem
--args="release_ver=0.0.18"

Change-Id: Ifc4b02aa647c4cca70fb4b3b7c84469a404df51f
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/494286
Reviewed-by: Wai-Hong Tam <waihong@google.com>