audio_CrasSanity: Remove dependency on network
We want to make this test simple and clear on what issue to catch.
Instead of fetch and play audio/video files, open a local url and
execute javascript to generate sound so this test won't fail on
network problems. This also makes the test run faster.

BUG=chromium:717823
TEST=test_that -b squawks audio_CrasSanity

Change-Id: I659957742b9942277856aacccd3c3f1c6e33d223
Reviewed-on: https://chromium-review.googlesource.com/494408
Commit-Ready: Qiang Xu <warx@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>