[autotest] Force host_scheduler to use metadata_reporter
metadata_reporter uploads metadata in bulk and in a separate thread.
This change allows host_scheduler to use metadata_reporter if it's
running. If the reporter is not running, it will still upload the data
using UDP, which may lead to package lost.

Also replace the email alert with metrics data.

BUG=None
TEST=run host_scheduler locally and run test, confirm the data is
uploaded.
Change-Id: I6eb173c239ad4043b701c8964b1c422b2afc10b7
Reviewed-on: https://chromium-review.googlesource.com/502716
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Dan Shi <dshi@google.com>