[autotest] Clean up cached property implementation
Use the non-data descriptor behavior of looking up instance attributes
before class attributes to cache the value.  This save an extra dict
and the danger of overwriting the instances _property_cache if it is
already using it, and a few extra attribute lookups when accessed.

This is a pretty standard cached property template.

BUG=None
TEST=Run unittests

Change-Id: I16cabcfa942baca673da80ccb92b297a64e9aca0
Reviewed-on: https://chromium-review.googlesource.com/503548
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>