packager: Compare checksum before skipping tar in upload
Earlier, we solely checked if a tarball was present before skipping
the tarring of the autotest tarball. But this tarball might be stale.
So, we additionally compare the checksum of the tarball against what
checksum the pacakger dictionary has. If there is a mismatch,
then we perform the tar again, and upload the new tarball (and update
the corresponding checksum).

This is a speculative patch. There doesn't seem to be a way to reproduce
the Bug in question locally, and so this CL is being pushed.

At the very least it should not lead to any further degradation in
performance or functionality.

BUG=chromium:716151
TEST='emerge-veyron_minnie chromeos-base/autotest-tests-wimax'

Change-Id: I20212e3a0c7159da0716b641a4573d00761584a9
Reviewed-on: https://chromium-review.googlesource.com/507008
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>