[moblab] Use credentials file from the bucket if available.
If the partner bucket has a credentials file - use that file and
enable the cloud notifications.

Fix issue with the command to restart after the wizard, in some
cases restart was not a good option as the service had not started
stop and start solves that issue.

Cleanup, only validate the boto key one time, since it is now a
more expensive operation.

TEST=unit tests, manually tested running GTS on local moblab
BUG=chromium:723863

Change-Id: I88e6d6901227fa667299a7c973d4f9276aceaf96
Reviewed-on: https://chromium-review.googlesource.com/508157
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>