Dump JMI not-aggregated dataset in a json file.
While we can use aggregated files to determine performance regressions,
we want to plot and analyse the collected JMI data in dedicated Google
dashboards. This way, we can collect more signals about performance
regressions.

The CL adds a method that collects interesting metrics and dumps them in
a raw format into a JSON file. This data can later be uploaded to CNS
and be moved to CNS and fed into Google Dashboards. How to format the
file to be digestible by the dashboards will be addressed separately.

BUG=chromium:723198
TEST=Run performance tests on a buddy device to collect the data locally.

Change-Id: Idd6117e041a2f1c05fe01ab0f7968561429f1824
Reviewed-on: https://chromium-review.googlesource.com/512863
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>